requirements.txt
python-whois>=0.7.3
requests>=2.31.0
ipwhois>=1.1.0
certifi>=2023.7.22
python-dotenv>=1.0.0
tabulate>=0.9.0
python-dateutil>=2.8.2
